+++ AbstractSortAlgorithm.h
@@ -15,13 +15,13 @@ class AbstractSortAlgorithm
 		AbstractSortAlgorithm();
 		AbstractSortAlgorithm(const AbstractSortAlgorithm& orig);
 		virtual ~AbstractSortAlgorithm();
-		virtual void sort(std::vector<int>&) = 0;
+		virtual void sort(std::vector<long long int>&) = 0;
 		virtual void print();
 		virtual void printRunningTime(std::string&) = 0;
 		virtual void exportRunningTime(std::string&, std::string&); /* common method to export running times to a txt file. */
 		virtual void exportSortedData(std::string prefix, std::string& file_name);
 	protected:
-		std::vector<int> data;
+		std::vector<long long int> data;
 		clock_t running_time;
 		std::ofstream output_stream;
 };
