+++ BubbleSort.h
@@ -8,7 +8,7 @@ class BubbleSort : public AbstractSortAlgorithm
 public:
 	BubbleSort();
 	~BubbleSort();
-	void sort(std::vector<int>&) override;
+	void sort(std::vector<long long int>&) override;
 	void bubbleSort();
 	void print() override;
 	void printRunningTime(std::string&) override;
